kubernetes/src/rest-api/requirements-dev.txt
fastapi[standard]~=0.115.4     # For the fastapi shell command
click~=8.1.7
PyYAML~=6.0.2
pydantic~=2.9.2
uvicorn~=0.32.0